Move type related funcs to types file
